Site fully deployed, fixed bugs, updated interface
# Fixed Image loading on landing page,

This problem was noticed on the previous pull request, while it was not present on my local development enviroment, there was an issue with the resource files in my azure.

I located the problem down to azure not taking all of the resource files when publishing from visual studio.

# Updated the redirect upon failed login attempt,

instead of going to a third error page, it simply goes back to the landing page and allows user to try logging in again

# updated interface

Wakepal Logo in an alternate colour version is now present on the admin dashboard

Created a favicon.ico icon for the little icon shown on the tab in a browser, added it to both landing page and the dashboard.